Update names.yaml
Adding Amazigh language name. Since the app forces for zgh standard Moroccan Tamazight. Which is long, confusing to many and actually representing the language keyboard itself which is used for all Amazigh variants. 

As well as the word Tamazight is not English, but rather Amazigh (Amazigh people, Amazigh language). Tamazight is the romanized version of the feminine form of that word in! Amazigh.